refactor: structure decrypter directory
