+++ x/vaultdecryption/auth/aws_iam_auth.go
@@ -1,4 +1,4 @@
-package vaultdecryption
+package auth
 
 import (
 	"context"
